requirements.txt
# Drone Firefighting System - Dependencies
# Compatible with Python 3.11+ (tested on 3.13)

# Database
sqlalchemy>=2.0.35
alembic>=1.13.1

# Web Framework & Dashboard
flask==3.0.0
flask-socketio==5.3.5
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Networking
requests==2.31.0
python-socketio==5.10.0

# Configuration
pyyaml==6.0.1
python-dateutil==2.8.2

# Drone Communication (Optional - for hardware mode)
pymavlink>=2.4.41
dronekit>=2.9.2
pyserial>=3.5
future>=1.0.0  # Required for dronekit Python 3.13 compatibility

# Machine Learning (optional - basic fire detection uses thermal thresholding)
# tensorflow==2.15.0  # Uncomment if you need advanced ML models
opencv-python>=4.8.0
pillow>=10.0.0

# Geospatial
folium==0.15.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tzlocal==5.2

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
